Optimize LMS authorization UX
Optimize the LMS authorization flow by having the server provide a hint
to the frontend about whether authorization has already happened. If it
has not, clicking the "PDF from <LMS Name>" button will show the
authorization popup immediately instead of flashing the LMS file picker
dialog briefly first. If it has, the user will see the file picker
dialog immediately when they click that button.

It is possible that the authorization may have expired or been revoked
since the frontend app was loaded, in which case it will show the file
picker dialog briefly but then revert to the authorization dialog when
the call to list files fails.