Adjust placement and style of Cancel buttons
Swap the order of the Cancel and affirmative action buttons in dialogs to feel
more consistent with h and other web interfaces (GitHub in particular was a
reference).

Also change the background of the Cancel button to better match h and
visually differentiate it from the affirmative action button.

"Cancel" buttons in the h frontend have no background at all. In this
context however this had the effect of making the button text look
misaligned. With a visible background, the button edges can be made to
lign up with the content of the dialog above.