Implement a Table component
Implement a Table component as the basis for an improved file list in
the LMS file picker dialog.

This component has selection behavior which is more like a classic file
dialog UI where the user has to click once to select an item or double
click to select and use.

It also implements basic keyboard navigation.